feat(exam): 基本完成考试管理模块

+++ exam-system-online/src/main/java/edu/jlu/intellilearnhub/server/entity/ExamRecord.java
@@ -22,7 +22,7 @@ public class ExamRecord extends BaseEntity {
 
     @Schema(description = "试卷ID，关联的考试试卷", 
             example = "1")
-    private Integer examId; // 试卷ID
+    private Long examId; // 试卷ID
 
     @Schema(description = "考生姓名", 
             example = "张三")
